+++ src/main/java/com/umc/finly/domain/member/exception/code/MemberErrorCode.java
@@ -1,4 +1,4 @@
-package com.umc.finly.domain.member.exception;
+package com.umc.finly.domain.member.exception.code;
 
 import com.umc.finly.global.apiPayload.response.BaseCode;
 import lombok.AllArgsConstructor;
